restructure opdyn model
