+++ examples/opdyn/mod.rs
@@ -1,0 +1,4 @@
+mod ent;
+mod sim;
+pub use self::ent::{Opinion, Person, Media};
+pub use self::sim::{OpinionDynamicsSim, World, State};
